tweaks to interpolation to work independent of current game speed; added debug renderer to show entity ids; added graph for render time
Former-commit-id: 9554c4eb32c2d46861c1516cb669a42849f61115

+++ EngineController/AbstractController.cs
@@ -29,7 +29,7 @@ namespace Engine.Controller
         /// The current 'load', i.e. how much of the available time is actually
         /// needed to perform an update.
         /// </summary>
-        public abstract double CurrentLoad { get; }
+        public abstract float CurrentLoad { get; }
 
         #endregion
